requirements.txt
numpy
trimesh
scipy
cadquery
cq-warehouse
cq-gears
flask
pythonocc-core
